server/requirements.txt
flask
flask-cors
numpy
scikit-learn  # if needed by your model
python-dotenv
fastapi
uvicorn
python-dotenv
google-generativeai
